Cambios en interfaz y precreacion de Factura - Se crearon nuevos estilos para la interfaz - Creacion de Controlador factura.

+++ Models/DetalleFactura.cs
@@ -9,7 +9,7 @@ public class DetalleFactura
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public int Codigo { get; set; }
     [Required]
-    public double DescuentoProducto { get; set; }
+    public decimal DescuentoProducto { get; set; }
     [Required]
     public int Cantidad { get; set; }
     [Required]
